# contracts/user.openapi.yaml
openapi: 3.1.0
info:
  title: Integra — user-service
  version: 2.1.0
  description: |
    Perfil, formação, vínculo institucional e o grafo de seguidores.

    ## A mudança de modelo da v2

    Até a v1 havia **uma** noção de afiliação, obrigatória no cadastro, que ao
    mesmo tempo aparecia no perfil e concedia acesso aos posts internos da
    instituição. Isso misturava duas coisas de naturezas diferentes. A v2 as
    separa:

    **Formação declarada** — o que a pessoa diz. Opcional, sem nenhuma
    verificação, como no LinkedIn: dá para declarar uma faculdade onde nunca se
    estudou. Aparece no perfil e **não concede acesso a nada**.

    **Vínculo institucional** — o que a faculdade confirma. Nasce quando o aluno
    informa o CPF no perfil da faculdade e aquele CPF consta na lista que ela
    cadastrou. É o **único** mecanismo que concede visibilidade de posts
    `institucional` e `curso`.

    ## Duas cardinalidades que não são a mesma

    | | quantidade | efeito |
    | --- | --- | --- |
    | `Formacao` | 0..N, acumula | cosmético; o selo de verificada é permanente |
    | `Vinculo` | **0..1**, se move | concede visibilidade; um por vez |

    Trocar de faculdade encerra o vínculo anterior e abre o novo, mas **as duas
    formações continuam verificadas** — quem se formou na primeira realmente
    estudou lá, e apagar o selo reescreveria o passado. A consequência a não
    esquecer: **formação verificada sem vínculo ativo não concede nada.**
servers:
...
paths:
...
  /users/me:
    get:
      tags: [perfil]
      summary: Perfil do usuário autenticado
      operationId: getMe
      responses:
        '200':
          description: Perfil próprio, com contato, formações e vínculo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Perfil' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }

    patch:
      tags: [perfil]
      summary: Atualiza o próprio perfil
      operationId: updateMe
      description: |
        Campos omitidos ficam inalterados.

        **Não editáveis aqui:** `email` e `cpf`. O e-mail é credencial e pertence
        ao auth-service. O CPF é a chave que liga a conta às matrículas das
        faculdades — deixá-lo editável permitiria assumir a matrícula de outra
        pessoa trocando o próprio CPF pelo dela.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AtualizarPerfilRequest' }
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Perfil' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '409':
          description: Username já em uso por outra conta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: username_ja_existe, message: 'Username já existe' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /users/{userId}:
    get:
      tags: [perfil]
      summary: Perfil público de outro usuário
      operationId: getUser
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: |
            Perfil público. Omite `email`, `telefone` e **`cpf`** — o CPF nunca
            sai do serviço em resposta pública, nem parcialmente mascarado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PerfilPublico' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ───────────────────────────  formação  ───────────────────────────
  /users/me/formacoes:
    post:
      tags: [formação]
      summary: Declara uma formação
      operationId: declararFormacao
      description: |
        Sem nenhuma verificação, por decisão de produto. A formação nasce com
        `verificadaEm: null` e só ganha o selo pelo fluxo de vínculo.

        O curso precisa pertencer à universidade informada — não por
        desconfiança, e sim porque os dois campos vêm de listas encadeadas na
        tela, e um par impossível indica requisição malformada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [universidadeId, cursoId]
              properties:
                universidadeId: { type: string, format: uuid }
                cursoId: { type: string, format: uuid }
      responses:
        '201':
          description: Formação declarada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Formacao' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '409':
          description: Esta formação já consta no perfil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /users/me/formacoes/{formacaoId}:
    delete:
      tags: [formação]
      summary: Remove uma formação do perfil
      operationId: removerFormacao
      description: |
        Remove apenas formações **não verificadas**. Uma verificada responde
        `409`: o selo é afirmação da instituição, não do usuário, e deixá-lo
        apagável pelo perfil transformaria "verificado" em algo que o próprio
        interessado controla. Para sair do vínculo existe
        `DELETE /users/me/vinculo`, que é outra coisa.
      parameters:
        - { name: formacaoId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Removida }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '404':
          description: Formação não encontrada neste perfil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Formação verificada não pode ser removida pelo usuário
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: formacao_verificada
                message: 'Formações verificadas pela instituição não podem ser removidas'

  # ───────────────────────────  vínculo  ───────────────────────────
  /universidades/{universidadeId}/vinculo:
    post:
      tags: [vínculo]
      summary: Informa o CPF para criar o vínculo com a instituição
      operationId: criarVinculo
      description: |
        O fluxo por trás do item "inserir CPF" no menu do perfil da faculdade.

        **Duas conferências, nesta ordem:**

        1. o CPF enviado é o **da própria conta**;
        2. esse CPF consta nas matrículas que esta universidade cadastrou.

        A primeira é o que impede a escalada. O CPF não é segredo no Brasil —
        circula em vazamentos e cadastros de todo tipo — então conferir apenas
        contra a lista da faculdade transformaria o CPF em senha de acesso aos
        comunicados internos dela. Com a conferência contra a conta, quem souber
        o CPF de outra pessoa para no primeiro passo.

        **Efeitos:** encerra o vínculo ativo anterior, se houver; cria o novo; e
        marca a formação correspondente como verificada, **sem apagar** as
        verificações antigas.
      parameters:
        - { name: universidadeId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cpf]
              properties:
                cpf:
                  type: string
                  description: Com ou sem pontuação; normalizado para dígitos.
                  examples: ['123.456.789-00']
      responses:
        '201':
          description: Vínculo criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vinculo' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403':
          description: |
            O CPF enviado não é o desta conta. Mensagem deliberadamente igual à
            do `404` abaixo no que o usuário vê, para a resposta não revelar se
            um CPF consta na lista de uma instituição.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: vinculo_recusado
                message: 'Não encontramos esse CPF na lista desta instituição'
        '404':
          description: CPF não cadastrado por esta universidade
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: vinculo_recusado
                message: 'Não encontramos esse CPF na lista desta instituição'
        '422': { $ref: '#/components/responses/ValidationError' }
  /users/me/vinculo:
    get:
      tags: [vínculo]
      summary: Vínculo ativo do usuário autenticado
      operationId: getMeuVinculo
      responses:
        '200':
          description: O vínculo, ou `null` quando não há nenhum
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/Vinculo' }
                  - { type: 'null' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }

    delete:
      tags: [vínculo]
      summary: Encerra o próprio vínculo
      operationId: encerrarVinculo
      description: |
        O aluno sai quando quiser, sem depender da instituição. Sem isto, quem
        trocou de faculdade e não foi removido da lista antiga ficaria preso
        recebendo comunicado de onde não estuda mais.

        A **formação segue verificada** — encerrar o vínculo não desfaz o fato de
        ter estudado lá. Voltar a inserir o CPF recria o vínculo, se a matrícula
        ainda existir.
      responses:
        '204': { description: Encerrado (idempotente — sem vínculo também responde 204) }
        '401': { $ref: '#/components/responses/NaoAutenticado' }

  # ─────────────────────────  instituições  ─────────────────────────
...
  /users:
    get:
      tags: [busca]
      summary: Busca somente pessoas, paginada
      operationId: searchUsers
      description: |
        A lista completa quando o usuário abre o grupo "Pessoas" da busca.
        `ILIKE` com índice trigram, em vez do truque `isLessThan: '${termo}z'`
        do protótipo, que só casava prefixo.
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 2 } }
        - { name: universidadeId, in: query, schema: { type: string, format: uuid },
            description: 'Filtra por VÍNCULO ativo, não por formação declarada.' }
        - { name: cursoId, in: query, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Página de pessoas
          content:
            application/json:
              schema:
                type: object
                required: [itens]
                properties:
                  itens:
                    type: array
                    items: { $ref: '#/components/schemas/PerfilPublico' }
                  proximoCursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/NaoAutenticado' }

  # ────────────────────────────  seguir  ────────────────────────────
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    TipoConta:
      type: string
      enum: [aluno, faculdade, empresa]
      description: |
        Só `aluno` nasce por autocadastro. `faculdade` e `empresa` são
        provisionados à mão enquanto não há verificação institucional — e são os
        únicos tipos sem CPF.
    Universidade:
      type: object
      required: [id, nome, sigla]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string, examples: ['Faculdade de Tecnologia de Ribeirão Preto'] }
        sigla: { type: string, examples: ['FATEC RP'] }
        temConta:
          type: boolean
          description: |
            Se já existe conta institucional administrando esta universidade. O
            perfil de uma sem conta não tem posts — e a tela diz isso, em vez de
            mostrar um feed vazio sem explicação.
    Curso:
      type: object
      required: [id, nome]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string, examples: ['Análise e Desenvolvimento de Sistemas'] }
    Formacao:
      type: object
      required: [id, universidade, curso, verificadaEm]
      description: |
        Uma linha do currículo. `verificadaEm` não nulo é o selo — e o cliente
        mostra **apenas o selo**, sem rótulo algum nas não verificadas: a
        ausência já comunica o suficiente.
      properties:
        id: { type: string, format: uuid }
        universidade: { $ref: '#/components/schemas/Universidade' }
        curso: { $ref: '#/components/schemas/Curso' }
        verificadaEm:
          type: [string, 'null']
          format: date-time
          description: Quando a instituição confirmou. `null` = apenas declarada.
        criadoEm: { type: string, format: date-time }
    Vinculo:
      type: object
      required: [universidade, curso, criadoEm]
      description: |
        O laço ativo. **No máximo um por usuário**, e o único que concede
        visibilidade de posts `institucional` e `curso`.
      properties:
        universidade: { $ref: '#/components/schemas/Universidade' }
        curso: { $ref: '#/components/schemas/Curso' }
        criadoEm: { type: string, format: date-time }
...
    PerfilPublico:
      type: object
      required: [id, nomeCompleto, username, tipo, formacoes, criadoEm]
      properties:
        id: { type: string, format: uuid }
        nomeCompleto: { type: string }
        username: { type: string }
        tipo: { $ref: '#/components/schemas/TipoConta' }
        formacoes:
          type: array
          description: Currículo visível. Pode estar vazio.
          items: { $ref: '#/components/schemas/Formacao' }
        vinculo:
          description: |
            Instituição atual, quando há. Público de propósito: é o equivalente
            a "trabalha em" num perfil profissional.
          oneOf:
            - { $ref: '#/components/schemas/Vinculo' }
            - { type: 'null' }
        fotoUrl: { type: [string, 'null'], format: uri }
        bio: { type: [string, 'null'], maxLength: 280 }
        criadoEm: { type: string, format: date-time }
    Perfil:
      allOf:
        - $ref: '#/components/schemas/PerfilPublico'
        - type: object
          required: [email]
          description: |
            O próprio perfil. **`cpf` e `cnpj` aparecem somente aqui** — nenhum dos
            dois sai em resposta pública, nem mascarado.
          properties:
            email: { type: string, format: email }
            cpf:
              type: [string, 'null']
              description: 'Apenas dígitos. Nulo em contas faculdade e empresa.'
            cnpj:
              type: [string, 'null']
              description: |
                Apenas dígitos. Nulo em contas de aluno. Uma conta nunca tem CPF e
                CNPJ ao mesmo tempo.
            ativadaEm:
              type: [string, 'null']
              format: date-time
              description: |
                Quando a conta passou a poder agir. **Nulo significa pendente**, e
                só acontece em conta institucional: a de aluno nasce ativa.

                Pendente, a instituição entra, edita o perfil e vê a própria
                página — mas `POST` e `DELETE` em `/universidades/me/*` respondem
                `403`. O cliente usa este campo para mostrar o aviso de análise.
            telefone: { type: [string, 'null'] }
            alteradoEm: { type: [string, 'null'], format: date-time }
...
    AtualizarPerfilRequest:
      type: object
      minProperties: 1
      description: 'Envie só o que mudou. `email` e `cpf` não são editáveis.'
      properties:
        nomeCompleto: { type: string }
        username: { type: string, minLength: 3, pattern: '^[a-zA-Z0-9_]+$' }
        telefone: { type: string, pattern: '^\(?\d{2}\)?[\s-]?\d{4,5}-?\d{4}$' }
        bio: { type: [string, 'null'], maxLength: 280 }
        fotoUrl: { type: [string, 'null'], format: uri }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
  responses:
    NaoAutenticado:
      description: Token ausente, inválido ou expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: nao_autenticado, message: 'Sessão expirada. Faça login novamente.' }
...
    ValidationError:
      description: Falha de validação, com as mensagens do cliente
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            code: validation_error
            message: 'Verifique os campos destacados'
            fields:
              cpf: ['CPF inválido']
security:
  - bearerAuth: []